Move reasoner benchmarks to test/
