+++ test/benchmark/BenchmarkSmallIT.java
@@ -29,6 +29,10 @@ import com.vaticle.typedb.core.concept.thing.Relation;
 import com.vaticle.typedb.core.concept.type.EntityType;
 import com.vaticle.typedb.core.concept.type.RelationType;
 import com.vaticle.typedb.core.database.CoreDatabaseManager;
+import com.vaticle.typedb.core.test.benchmark.generation.DiagonalGraph;
+import com.vaticle.typedb.core.test.benchmark.generation.PathTreeGraph;
+import com.vaticle.typedb.core.test.benchmark.generation.TransitivityChainGraph;
+import com.vaticle.typedb.core.test.benchmark.generation.TransitivityMatrixGraph;
 import com.vaticle.typedb.core.test.integration.util.Util;
 import com.vaticle.typeql.lang.TypeQL;
 import com.vaticle.typeql.lang.query.TypeQLMatch;
