+++ test/benchmark/generation/TransitivityChainGraph.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.vaticle.typedb.core.reasoner.benchmark;
+package com.vaticle.typedb.core.test.benchmark.generation;
 
 import com.vaticle.typedb.core.TypeDB;
 import com.vaticle.typedb.core.common.parameters.Arguments;
@@ -25,22 +25,24 @@ import com.vaticle.typedb.core.concept.thing.Relation;
 import com.vaticle.typedb.core.concept.thing.Thing;
 import com.vaticle.typedb.core.concept.type.EntityType;
 import com.vaticle.typedb.core.concept.type.RelationType;
+import com.vaticle.typedb.core.concept.type.RoleType;
+import com.vaticle.typedb.core.test.benchmark.Util;
 
 @SuppressWarnings("CheckReturnValue")
-public class DiagonalGraph{
+public class TransitivityChainGraph {
 
     private final TypeDB.DatabaseManager dbm;
     private final String databaseName;
 
-    private static final String schemaFile = "test/integration/reasoner/benchmark/resources/diagonalTest.tql";
-    private static final Label key = Label.of("name");
+    private final static String schemaFile = "test/benchmark/resources/quadraticTransitivity.tql";
+    private final static Label key = Label.of("index");
 
-    public DiagonalGraph(TypeDB.DatabaseManager dbm, String dbName){
+    public TransitivityChainGraph(TypeDB.DatabaseManager dbm, String dbName){
         this.dbm = dbm;
         this.databaseName = dbName;
     }
 
-    public final void load(int n, int m) {
+    public final void load(int n) {
         try (TypeDB.Session session = dbm.session(databaseName, Arguments.Session.Type.SCHEMA)) {
             try (TypeDB.Transaction tx = session.transaction(Arguments.Transaction.Type.WRITE)) {
                 tx.query().define(Util.parseTQL(schemaFile).asDefine());
@@ -50,43 +52,32 @@ public class DiagonalGraph{
 
         try (TypeDB.Session session = dbm.session(databaseName, Arguments.Session.Type.DATA)) {
             try (TypeDB.Transaction tx = session.transaction(Arguments.Transaction.Type.WRITE)) {
-                buildExtensionalDB(n, m, tx);
+                buildExtensionalDB(n, tx);
                 tx.commit();
             }
         }
     }
 
-    protected void buildExtensionalDB(int n, int m, TypeDB.Transaction tx) {
-
-        EntityType entity1 = tx.concepts().getEntityType("entity1");
-        RelationType horizontal = tx.concepts().getRelationType("horizontal");
-        RelationType vertical = tx.concepts().getRelationType("vertical");
-        Thing[][] instanceIds = new Thing[n][m];
-        long inserts = 0;
-        for (int i = 0; i < n; i++) {
-            for (int j = 0; j < m; j++) {
-                instanceIds[i][j] = Util.putEntityWithResource(tx, "a" + i + "," + j, entity1, key);
-                inserts++;
-                if (inserts % 100 == 0) System.out.println("inst inserts: " + inserts);
-            }
+    protected void buildExtensionalDB(int n, TypeDB.Transaction tx){
+        EntityType aEntity = tx.concepts().getEntityType("a-entity");
+        RelationType Q = tx.concepts().getRelationType("Q");
+        RoleType qfrom = Q.getRelates("from");
+        RoleType qto = Q.getRelates("to");
+        Thing aInst = Util.putEntityWithResource(tx, "a", tx.concepts().getEntityType("entity2"), key);
+        Thing[] aInstanceIds = new Thing[n];
+        for(int i = 0 ; i < n ;i++) {
+            aInstanceIds[i] = Util.putEntityWithResource(tx, "a" + i, aEntity, key);
         }
 
-        for (int i = 0; i < n; i++) {
-            for (int j = 0; j < m; j++) {
-                if (i < n - 1) {
-                    Relation v = vertical.create();
-                    v.addPlayer(vertical.getRelates("from"), instanceIds[i][j]);
-                    v.addPlayer(vertical.getRelates("to"), instanceIds[i + 1][j]);
-                    inserts++;
-                }
-                if (j < m - 1) {
-                    Relation h = horizontal.create();
-                    h.addPlayer(horizontal.getRelates("from"), instanceIds[i][j]);
-                    h.addPlayer(horizontal.getRelates("to"), instanceIds[i][j + 1]);
-                    inserts++;
-                }
-                if (inserts % 100 == 0) System.out.println("rel inserts: " + inserts);
-            }
+        Relation q0 = Q.create();
+        q0.addPlayer(qfrom, aInst);
+        q0.addPlayer(qto, aInstanceIds[0]);
+
+        for(int i = 0 ; i < n - 1 ; i++) {
+            Relation q = Q.create();
+            q.addPlayer(qfrom, aInstanceIds[i]);
+            q.addPlayer(qto, aInstanceIds[i+1]);
         }
     }
+
 }
