+++ test/benchmark/generation/TransitivityMatrixGraph.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.vaticle.typedb.core.reasoner.benchmark;
+package com.vaticle.typedb.core.test.benchmark.generation;
 
 import com.vaticle.typedb.core.TypeDB;
 import com.vaticle.typedb.core.common.parameters.Arguments;
@@ -26,22 +26,25 @@ import com.vaticle.typedb.core.concept.thing.Thing;
 import com.vaticle.typedb.core.concept.type.EntityType;
 import com.vaticle.typedb.core.concept.type.RelationType;
 import com.vaticle.typedb.core.concept.type.RoleType;
+import com.vaticle.typedb.core.test.benchmark.Util;
+
 
 @SuppressWarnings("CheckReturnValue")
-public class TransitivityChainGraph {
+public abstract class TransitivityMatrixGraph {
 
     private final TypeDB.DatabaseManager dbm;
     private final String databaseName;
+    private final String schemaFile;
 
-    private final static String schemaFile = "test/integration/reasoner/benchmark/resources/quadraticTransitivity.tql";
     private final static Label key = Label.of("index");
 
-    public TransitivityChainGraph(TypeDB.DatabaseManager dbm, String dbName){
+    public TransitivityMatrixGraph(String schemaFile, TypeDB.DatabaseManager dbm, String dbName){
         this.dbm = dbm;
         this.databaseName = dbName;
+        this.schemaFile = schemaFile;
     }
 
-    public final void load(int n) {
+    public final void load(int n, int m) {
         try (TypeDB.Session session = dbm.session(databaseName, Arguments.Session.Type.SCHEMA)) {
             try (TypeDB.Transaction tx = session.transaction(Arguments.Transaction.Type.WRITE)) {
                 tx.query().define(Util.parseTQL(schemaFile).asDefine());
@@ -51,32 +54,61 @@ public class TransitivityChainGraph {
 
         try (TypeDB.Session session = dbm.session(databaseName, Arguments.Session.Type.DATA)) {
             try (TypeDB.Transaction tx = session.transaction(Arguments.Transaction.Type.WRITE)) {
-                buildExtensionalDB(n, tx);
+                buildExtensionalDB(n, m, tx);
                 tx.commit();
             }
         }
     }
 
-    protected void buildExtensionalDB(int n, TypeDB.Transaction tx){
+    protected void buildExtensionalDB(int n, int m, TypeDB.Transaction tx){
+
         EntityType aEntity = tx.concepts().getEntityType("a-entity");
         RelationType Q = tx.concepts().getRelationType("Q");
-        RoleType qfrom = Q.getRelates("from");
-        RoleType qto = Q.getRelates("to");
+        RoleType Qfrom = Q.getRelates("from");
+        RoleType Qto = Q.getRelates("to");
+
+
+        Thing[][] aInstancesIds = new Thing[n+1][m+1];
         Thing aInst = Util.putEntityWithResource(tx, "a", tx.concepts().getEntityType("entity2"), key);
-        Thing[] aInstanceIds = new Thing[n];
-        for(int i = 0 ; i < n ;i++) {
-            aInstanceIds[i] = Util.putEntityWithResource(tx, "a" + i, aEntity, key);
+        for(int i = 1 ; i <= n ;i++) {
+            for (int j = 1; j <= m; j++) {
+                aInstancesIds[i][j] = Util.putEntityWithResource(tx, "a" + i + "," + j, aEntity, key);
+            }
         }
 
-        Relation q0 = Q.create();
-        q0.addPlayer(qfrom, aInst);
-        q0.addPlayer(qto, aInstanceIds[0]);
+        Relation rel = Q.create();
+        rel.addPlayer(Qfrom, aInst);
+        rel.addPlayer(Qto, aInstancesIds[1][1]);
 
-        for(int i = 0 ; i < n - 1 ; i++) {
-            Relation q = Q.create();
-            q.addPlayer(qfrom, aInstanceIds[i]);
-            q.addPlayer(qto, aInstanceIds[i+1]);
+        for(int i = 1 ; i <= n ; i++) {
+            for (int j = 1; j <= m; j++) {
+                if ( i < n ) {
+                    Relation q = Q.create();
+                    q.addPlayer(Qfrom, aInstancesIds[i][j]);
+                    q.addPlayer(Qto, aInstancesIds[i][j]);
+                }
+                if (j < m) {
+                    Relation q = Q.create();
+                    q.addPlayer(Qfrom, aInstancesIds[i][j]);
+                    q.addPlayer(Qto, aInstancesIds[i][j+1]);
+                }
+            }
         }
     }
 
+    public static class Linear extends TransitivityMatrixGraph {
+        private static final String schemaFile = "test/benchmark/resources/linearTransitivity.tql";
+
+        public Linear(TypeDB.DatabaseManager dbm, String dbName) {
+            super(schemaFile, dbm, dbName);
+        }
+    }
+
+    public static class Quadratic extends TransitivityMatrixGraph {
+        private static final String schemaFile = "test/benchmark/resources/linearTransitivity.tql";
+
+        public Quadratic(TypeDB.DatabaseManager dbm, String dbName) {
+            super(schemaFile, dbm, dbName);
+        }
+    }
 }
