+++ test/benchmark/resources/diagonalTest.tql
