+++ test/benchmark/resources/linearTransitivity.tql
