+++ test/benchmark/resources/quadraticTransitivity.tql
